requirements.txt
# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
pyodbc>=5.0.0

# Environment
python-dotenv>=1.0.0

# Hyperliquid
hyperliquid-python-sdk>=0.20.0

# Data Analysis (versioni che hanno wheel pre-compilati)
pandas>=2.2.0
numpy>=2.0.0
ta>=0.11.0

# HTTP & Async
aiohttp>=3.9.0
requests>=2.31.0

# LLM
anthropic>=0.39.0

# Utilities
pydantic>=2.5.0
python-dateutil>=2.8.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0